streamline API responses (#9)
The goal of this is to make the API more compliant with REST best practices.
This commit adds the API version to the request and removes it from the response.
The endpoint locations have been unified (i.e. always `/streams/` instead of
sometimes `/stream/` and sometimes `/streams/`.
Responses to requests that do not contain a stream name no longer contain urls.
The app has been removed from requests for stream information.
multiple urls can now be returned. I.e. that we now return both flv and rtmp link.

Co-authored-by: Jonas <cpp@zom.bi>